comments

+++ src/main/java/myapp/SzakdolgozatBE/episode/Episode.java
@@ -105,7 +105,7 @@ public class Episode implements Serializable {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
